moving datasources to a separate folder
