+++ src/Events.ts
@@ -1,5 +1,5 @@
+import {HttpClientFactory, IODataSource} from './datasources/IODataSource'
 import {HttpClient, withoutRecorder} from './HttpClient'
-import {HttpClientFactory, IODataSource} from './IODataSource'
 
 const eventRoute = (route: string) => `/events/${route}`
