+++ src/Registry.ts
@@ -5,8 +5,8 @@ import {extract} from 'tar-fs'
 import {createGunzip, ZlibOptions} from 'zlib'
 
 import {DEFAULT_WORKSPACE} from './constants'
+import {HttpClientFactory, IODataSource} from './datasources/IODataSource'
 import {HttpClient} from './HttpClient'
-import {HttpClientFactory, IODataSource} from './IODataSource'
 import {AppBundlePublished, AppFilesList, AppManifest} from './responses'
 
 const EMPTY_OBJECT = {}
