+++ src/index.ts
@@ -14,4 +14,4 @@ export * from './Billing'
 export * from './caches'
 export * from './MetricsAccumulator'
 export * from './utils'
-export * from './IODataSource'
+export * from './datasources'
